Simplify to use vite for building playground library
