+++ packages/playground/package.json
@@ -43,16 +43,16 @@
       "types": "./dist/src/react/viewers/index.d.ts",
       "default": "./dist/react/viewers/index.js"
     },
-    "./style.css": "./dist/index.css",
-    "./styles.css": "./dist/index.css"
+    "./style.css": "./dist/style.css",
+    "./styles.css": "./dist/style.css"
   },
   "engines": {
     "node": ">=18.0.0"
   },
   "scripts": {
     "clean": "rimraf ./dist ./dist-dev ./temp ./typespecContents.json",
-    "build": "rollup --config rollup.config.ts --configPlugin typescript --failAfterWarnings 2>&1",
-    "watch": "rollup --config rollup.config.ts --configPlugin typescript --watch",
+    "build": "vite build",
+    "watch": "vite build --watch",
     "copy-css": "copyfiles -u 1 src/**/*.module.css dist/src",
     "preview": "npm run build && vite preview",
     "start": "vite",
@@ -91,10 +91,6 @@
   "devDependencies": {
     "@babel/core": "^7.24.5",
     "@playwright/test": "^1.44.0",
-    "@rollup/plugin-commonjs": "~25.0.7",
-    "@rollup/plugin-json": "~6.1.0",
-    "@rollup/plugin-node-resolve": "~15.2.3",
-    "@rollup/plugin-typescript": "~11.1.6",
     "@types/debounce": "~1.2.4",
     "@types/node": "~18.11.19",
     "@types/react": "~18.3.2",
@@ -106,10 +102,8 @@
     "copyfiles": "~2.4.1",
     "cross-env": "~7.0.3",
     "rimraf": "~5.0.7",
-    "rollup": "~4.17.2",
-    "rollup-plugin-postcss": "~4.0.2",
-    "rollup-plugin-visualizer": "~5.12.0",
     "typescript": "~5.4.5",
-    "vite": "^5.2.11"
+    "vite": "^5.2.11",
+    "vite-plugin-dts": "^3.9.1"
   }
 }
