+++ packages/playground/src/index.ts
@@ -1,5 +1,5 @@
 export { createBrowserHost } from "./browser-host.js";
 export { registerMonacoDefaultWorkersForVite } from "./monaco-worker.js";
 export { registerMonacoLanguage } from "./services.js";
-export { StateStorage, UrlStateStorage, createUrlStateStorage } from "./state-storage.js";
+export { createUrlStateStorage, type StateStorage, type UrlStateStorage } from "./state-storage.js";
 export type { BrowserHost, PlaygroundSample } from "./types.js";
