+++ packages/playground/src/react/index.ts
@@ -2,12 +2,12 @@ export { usePlaygroundContext } from "./context/index.js";
 export {
   Footer,
   FooterItem,
-  FooterItemProps,
-  FooterProps,
   FooterVersionItem,
-  FooterVersionItemProps,
-  VersionSelectorProps,
-  VersionSelectorVersion,
+  type FooterItemProps,
+  type FooterProps,
+  type FooterVersionItemProps,
+  type VersionSelectorProps,
+  type VersionSelectorVersion,
 } from "./footer/index.js";
 export { Playground } from "./playground.js";
 export type { PlaygroundProps, PlaygroundSaveData } from "./playground.js";
